pyproject.toml
[tool.poetry]
name = "locallm"
version = "0.1.0"
description = "Local LLM server with FastAPI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
    { include = "websrc" }
]

[tool.poetry.dependencies]
python = "^3.10" 
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
opentelemetry-sdk = "^1.28.1"
opentelemetry-instrumentation = "^0.49b1"
opentelemetry-instrumentation-fastapi = "^0.49b1"
opentelemetry-instrumentation-logging = "^0.49b1"
opentelemetry-instrumentation-requests = "^0.49b1"
pydantic-settings = "^2.6.1"
pydantic = "^2.9.2"
aiocache = "^0.12.3"
opentelemetry-exporter-otlp = "^1.28.1"
jinja2 = "^3.1.4"
opentelemetry-distro = "^0.49b1"
python-dotenv = "^1.0.1"
opentelemetry-exporter-jaeger-thrift = "^1.21.0"
python-multipart = "^0.0.17"
pytest = "^8.3.3"
python-json-logger = "^2.0.7"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
redis = "^5.0.1"
tenacity = "^9.0.0"
sqlalchemy-orm = "^1.2.10"
torch = "^2.5.1"
transformers = "^4.46.3"
sentencepiece = "^0.2.0"
accelerate = "^1.1.1"
aiofiles = "^24.1.0"
cryptography-fernet = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"